# docs/api/openapi.yaml
components:
  schemas:
    BacktestRequestBody:
      properties:
        as_of:
          format: date-time
          title: As Of
          type: string
        cost_assumptions:
          additionalProperties: true
          title: Cost Assumptions
          type: object
        dataset_snapshot_ids:
          items:
            type: string
          minItems: 1
          title: Dataset Snapshot Ids
          type: array
        formula_version:
          minLength: 1
          title: Formula Version
          type: string
        git_sha:
          minLength: 1
          title: Git Sha
          type: string
        model_run_id:
          minLength: 1
          title: Model Run Id
          type: string
        pipeline_inputs:
          additionalProperties: true
          title: Pipeline Inputs
          type: object
        strategy_id:
          minLength: 1
          title: Strategy Id
          type: string
        validation_protocol:
          minLength: 1
          title: Validation Protocol
          type: string
      required:
      - strategy_id
      - dataset_snapshot_ids
      - validation_protocol
      - cost_assumptions
      - formula_version
      - model_run_id
      - git_sha
      - as_of
      title: BacktestRequestBody
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: AI Infrastructure Fund API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /internal/backtests:
    post:
      operationId: create_backtest_internal_backtests_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestRequestBody'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Backtest
  /internal/backtests/recent:
    get:
      operationId: recent_backtests_internal_backtests_recent_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Recent Backtests
  /internal/backtests/{request_id}:
    get:
      operationId: get_backtest_internal_backtests__request_id__get
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Backtest
...
  /internal/events/by-run/{run_id}:
    get:
      operationId: for_run_events_internal_events_by_run__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: For Run Events
  /internal/events/recent:
    get:
      operationId: recent_events_internal_events_recent_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: since
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Since
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Recent Events
...
  /internal/shadow-portfolio/drift:
    get:
      operationId: shadow_drift_internal_shadow_portfolio_drift_get
      parameters:
      - in: query
        name: as_of
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: As Of
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Shadow Drift
  /internal/shadow-portfolio/simulation:
    get:
      operationId: shadow_simulation_internal_shadow_portfolio_simulation_get
      parameters:
      - in: query
        name: as_of
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: As Of
      - in: query
        name: horizon_days
        required: false
        schema:
          default: 30
          maximum: 365
          minimum: 1
          title: Horizon Days
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Shadow Simulation
...